requirements.txt
streamlit
wikipedia-api
sentence-transformers
langchain
faiss-cpu
transformers
